Make core aspects friendly to load time weaving
Apps startup quicker if you weave the aspects at compile time, or
via an agent. These changes show how to modify the @Aspects in
Sleuth and make them accessible to AspectJ. TL;DR you need
a default constructor and a @Bean method that sets the properties
on the Aspects.aspectOf().